adds splink and removes kanoniv since it doesnt seem to work that well. Splink job takes tables from delta and writes to a temp duckdb, runs splink, and exports link to a parquet.

+++ src/data_integration_pipeline/jobs/upload_bronze.py
@@ -12,6 +12,8 @@ class UploadBronzeJob:
     def run(self):
         for file in os.listdir(TESTS_DATA):
             local_file_path = os.path.join(TESTS_DATA, file)
+            if not os.path.isfile(local_file_path):
+                continue
             parent_folder = Path(file).stem
             s3_path = os.path.join(BRONZE_DATA_FOLDER, parent_folder, file)
             upload_status = self.s3_client.upload_file(local_path=local_file_path, s3_path=s3_path)
